requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
opencv-python-headless>=4.8
numpy>=1.24
Pillow>=10.0
